requirements.txt
Flask==0.10.1
Flask-OAuth==0.12
Flask-WTF==0.9.3
Jinja2==2.7.1
MarkupSafe==0.18
WTForms==1.0.5
Werkzeug==0.9.4
argparse==1.2.1
distribute==0.6.34
flask-mongoengine==0.7.0
gevent==0.13.8
gevent-socketio==0.3.5-rc2
gevent-websocket==0.3.6
greenlet==0.4.1
gunicorn==0.16.1
httplib2==0.8
itsdangerous==0.23
mongoengine==0.8.4
oauth2==1.5.211
pymongo==2.6.2
redis==2.8.0
wsgiref==0.1.2
